Changing from magento1 and magento2 being hosted in magescotch to hosting it in the magemalt base box

+++ public/magento2/dev/tests/integration/testsuite/Magento/Checkout/_files/quote_with_items_saved_rollback.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright © 2015 Magento. All rights reserved.
+ * Copyright © 2016 Magento. All rights reserved.
  * See COPYING.txt for license details.
  */
 $objectManager = \Magento\TestFramework\Helper\Bootstrap::getObjectManager();
@@ -9,14 +9,17 @@ $registry->unregister('isSecureArea');
 $registry->register('isSecureArea', true);
 $quote = $objectManager->create('Magento\Quote\Model\Quote');
 $quote->load('test_order_item_with_items', 'reserved_order_id');
+$quoteId = $quote->getId();
+if ($quote->getId()) {
+    $quote->delete();
+}
 $product = $objectManager->create('Magento\Catalog\Model\Product');
-foreach ($quote->getAllItems() as $item) {
-    $sku = $item->getSku();
-    $product->load($product->getIdBySku($sku));
-    if ($product->getId()) {
-        $product->delete();
-    }
-};
-$quote->delete();
+$product->load($product->getIdBySku('simple_one'));
+if ($product->getId()) {
+    $product->delete();
+}
+
 $registry->unregister('isSecureArea');
 $registry->register('isSecureArea', false);
+
+require __DIR__ . '/../../Checkout/_files/quote_with_address_rollback.php';
